Controller Added

+++ cars24/src/main/java/com/cg/cars/dao/IPaymentRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.cg.cars.entities.Payment;
 
 @Repository
-public interface IPaymentRepository extends JpaRepository<Payment,Integer> {
+public interface IPaymentRepository extends JpaRepository<Payment,Long> {
 
 }
